month-1/week-1/day-1/pyproject.toml
[project]
name = "day-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]
